// EPS-Client/src/app/modules/api/model/registerRequest.ts
/**
 * EPS API v1.0
 * API für das EPS
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RegisterRequest { 
    email: string;
    password: string;
    firstname: string;
    lastname: string;
    dateOfBirth: Date;
}